re-factor `store` commands
copyright year update for new files
